Moving majority_location query
