+++ flowmachine/flowmachine/core/server/query_schemas/location_visits.py
@@ -27,3 +27,5 @@ class LocationVisitsSchema(BaseSchema):
         fields.Nested(DailyLocationSchema),
         validate=OneOf([DailyLocationSchema, ModalLocationSchema]),
     )
+
+    __model__ = LocationVisitsExposed
